styled 컴포넌트 적용 디렉토리 구조변경
